backend/requirements.txt
Flask>=3.1,<3.2
Flask-SQLAlchemy>=3.1,<3.2
Flask-Cors>=6.0,<6.1
Flask-Bcrypt>=1.0,<1.1
Flask-Login>=0.6,<0.7
Flask-WTF>=1.2,<1.3
Flask-Migrate>=4.1,<4.2
Flask-Limiter>=3.8,<4.0
python-dotenv>=1.1,<1.2
google-generativeai>=0.7,<0.8
requests>=2.32,<3
pytest>=9,<10
gunicorn>=23,<24
psycopg2-binary>=2.9,<3
